fix: ReviewPostBox 분리
